requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ai21=2.0.5=pypi_0
ai21-tokenizer=0.3.11=pypi_0
aiohttp=3.9.3=pypi_0
aiolimiter=1.1.0=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
asgiref=3.7.2=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
bce-python-sdk=0.9.5=pypi_0
bcrypt=4.1.2=pypi_0
beautifulsoup4=4.12.3=pypi_0
bs4=0.0.2=pypi_0
build=1.1.1=pypi_0
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2023.12.12=haa95532_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.4.24=pypi_0
clevercsv=0.8.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
dataclasses-json=0.6.4=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
django=5.0.3=pypi_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.110.0=pypi_0
filelock=3.13.1=pypi_0
flatbuffers=24.3.7=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
future=1.0.0=pypi_0
google-auth=2.28.1=pypi_0
googleapis-common-protos=1.62.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.62.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.4=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.21.4=pypi_0
humanfriendly=10.0=pypi_0
idna=3.6=pypi_0
ijson=3.2.3=pypi_0
importlib-metadata=6.11.0=pypi_0
importlib-resources=6.1.3=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
kubernetes=29.0.0=pypi_0
langchain=0.1.11=pypi_0
langchain-ai21=0.1.0=pypi_0
langchain-community=0.0.27=pypi_0
langchain-core=0.1.30=pypi_0
langchain-openai=0.0.8=pypi_0
langchain-text-splitters=0.0.1=pypi_0
langchainhub=0.1.15=pypi_0
langsmith=0.1.22=pypi_0
libffi=3.4.4=hd77b12b_0
markdown-it-py=3.0.0=pypi_0
marshmallow=3.21.1=pypi_0
mdurl=0.1.2=pypi_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.0.0=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.17.1=pypi_0
openai=1.13.3=pypi_0
openssl=3.0.13=h2bbff1b_0
opentelemetry-api=1.23.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.23.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.23.0=pypi_0
opentelemetry-instrumentation=0.44b0=pypi_0
opentelemetry-instrumentation-asgi=0.44b0=pypi_0
opentelemetry-instrumentation-fastapi=0.44b0=pypi_0
opentelemetry-proto=1.23.0=pypi_0
opentelemetry-sdk=1.23.0=pypi_0
opentelemetry-semantic-conventions=0.44b0=pypi_0
opentelemetry-util-http=0.44b0=pypi_0
orjson=3.9.15=pypi_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
pip=23.3.1=py310haa95532_0
posthog=3.5.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.3=pypi_0
pulsar-client=3.4.0=pypi_0
pyarrow=15.0.1=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.6.3=pypi_0
pydantic-core=2.16.3=pypi_0
pygments=2.17.2=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.0.0=pypi_0
pyreadline3=3.4.1=pypi_0
python=3.10.13=he1021f5_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pyyaml=6.0.1=pypi_0
qianfan=0.3.4=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.7.1=pypi_0
rsa=4.9=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.2.2=py310haa95532_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.28=pypi_0
sqlite=3.41.2=h2bbff1b_0
sqlparse=0.4.4=pypi_0
starlette=0.36.3=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tiktoken=0.6.0=pypi_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.15.2=pypi_0
tomli=2.0.1=pypi_0
tqdm=4.66.2=pypi_0
typer=0.9.0=pypi_0
types-requests=2.31.0.20240218=pypi_0
typing-extensions=4.10.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.27.1=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchfiles=0.21.0=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.7.0=pypi_0
websockets=12.0=pypi_0
wheel=0.41.2=py310haa95532_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h8cc25b3_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h8cc25b3_0
